Back to original and add a one more step to do the `js` task
I noticed the folder `lib` is not at the repository at master branch and when runned `npm run build` it was expecting to already have this folder so
it'd copy its files to `build` folder. To solve it, it's needed to execute `js` task before `build-js-lib` so it'll copy files from `node_modules` to `src` before send them to `build` folder.